+++ api/main.py
@@ -23,8 +23,14 @@ from contextlib import asynccontextmanager
 from typing import Optional
 
 import torch
-from fastapi import FastAPI, HTTPException
+import pandas as pd
+import numpy as np
+import unicodedata
+import re
+from functools import lru_cache
+from fastapi import FastAPI, HTTPException, Depends, status
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from pydantic import BaseModel
 
 # ── Logging ────────────────────────────────────────────────────────────────────
@@ -52,6 +58,33 @@ except ImportError:
     TOXICITY_THRESHOLD = 0.82
     HF_QWEN_REPO       = "alfersal/qwen2.5-7b-deporte"
     HF_TOXICITY_REPO   = "alfersal/toxicity-deporte-es"
+
+# ── Seguridad y Autenticación ────────────────────────────────────────────────
+USERS_DB = {
+    "admin": {
+        "username": "admin",
+        "password": "1234",
+        "role": "administrator"
+    }
+}
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/v1/token")
+
+def get_current_user(token: str = Depends(oauth2_scheme)):
+    if not token.startswith("fake-token-"):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    username = token.replace("fake-token-", "")
+    if username not in USERS_DB:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="User not found",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    return USERS_DB[username]
 
 # ── Estado compartido de la app ────────────────────────────────────────────────
 class ModelState:
@@ -178,6 +211,153 @@ class HealthResponse(BaseModel):
     error: Optional[str] = None
 
 
+# ── Utilidades de Procesamiento de Datos ──────────────────────────────────────
+def repair_mojibake(text: str) -> str:
+    """Repara textos UTF-8 mal decodificados como latin-1 cuando es posible."""
+    if pd.isna(text): return ""
+    clean = str(text).replace("\ufeff", "").replace("ï»¿", "")
+    try:
+        repaired = clean.encode("latin1").decode("utf-8")
+        bad_before = clean.count("Ã") + clean.count("ï")
+        bad_after = repaired.count("Ã") + repaired.count("ï")
+        if bad_after <= bad_before:
+            clean = repaired
+    except (UnicodeEncodeError, UnicodeDecodeError):
+        pass
+    return clean.strip()
+
+def normalize_column_name(name: str) -> str:
+    """Normaliza nombres de columnas eliminando BOM, tildes, espacios y guiones."""
+    clean = repair_mojibake(name).lower()
+    clean = unicodedata.normalize("NFKD", clean)
+    clean = "".join(c for c in clean if not unicodedata.combining(c))
+    return re.sub(r'[^a-z0-t0-9]', '', clean)
+
+def normalize_for_match(text: str) -> str:
+    """Normalización ultra-agresiva para cruzar datos (CCAA, Federaciones)."""
+    clean = repair_mojibake(text).lower()
+    clean = unicodedata.normalize("NFKD", clean)
+    clean = "".join(c for c in clean if not unicodedata.combining(c))
+    return re.sub(r'[^a-z]', '', clean)
+
+def coalesce_normalized_columns(df: pd.DataFrame, canonical_map: dict[str, list[str]]) -> pd.DataFrame:
+    """Crea columnas canónicas combinando variantes equivalentes por nombre."""
+    result = df.copy()
+    normalized_lookup: dict[str, list[str]] = {}
+    for col in result.columns:
+        normalized_lookup.setdefault(normalize_column_name(col), []).append(col)
+
+    for canonical_name, aliases in canonical_map.items():
+        candidate_cols: list[str] = []
+        for alias in aliases:
+            alias_key = normalize_column_name(alias)
+            candidate_cols.extend(normalized_lookup.get(alias_key, []))
+        
+        # También buscar si el propio nombre canónico normalizado coincide con alguna columna
+        can_key = normalize_column_name(canonical_name)
+        candidate_cols.extend(normalized_lookup.get(can_key, []))
+            
+        if not candidate_cols:
+            continue
+        unique_candidates = list(dict.fromkeys(candidate_cols))
+        result[canonical_name] = result[unique_candidates].bfill(axis=1).iloc[:, 0]
+
+    return result
+
+def parse_spanish_number(value):
+    """Convierte números en formato español a float."""
+    if pd.isna(value):
+        return np.nan
+    text = str(value).strip().replace("\xa0", "").replace(" ", "")
+    if not text or text == "..":
+        return np.nan
+    if "," in text:
+        text = text.replace(".", "").replace(",", ".")
+    elif text.count(".") > 1:
+        text = text.replace(".", "")
+    elif text.count(".") == 1:
+        left, right = text.split(".")
+        if right.isdigit() and len(right) == 3 and left.isdigit():
+            text = left + right
+    try:
+        return float(text)
+    except ValueError:
+        return np.nan
+
+@lru_cache(maxsize=32)
+def build_home_data(year: int) -> pd.DataFrame:
+    """Carga y une federados.parquet + gasto.parquet para el año indicado."""
+    log.info(f"📊 build_home_data({year}): Iniciando carga...")
+    base_dir = os.path.join(os.path.dirname(os.path.dirname(__file__)), "data", "processed")
+    fed_path = os.path.join(base_dir, "federados.parquet")
+    gas_path = os.path.join(base_dir, "gasto.parquet")
+    
+    if not os.path.exists(fed_path) or not os.path.exists(gas_path):
+        log.error(f"Archivos no encontrados: {fed_path} o {gas_path}")
+        raise FileNotFoundError("Parquet files not found in data/processed/")
+
+    EXCLUDED_CCAA_RAW = {'TOTAL', 'Sin territorializar', 'Ceuta', 'Melilla'}
+    EXCLUDED_KEYS = {normalize_for_match(c) for c in EXCLUDED_CCAA_RAW}
+
+    # 1. Cargar y filtrar por año inmediatamente para reducir memoria
+    yr_str = str(year)
+    fed_raw = pd.read_parquet(fed_path)
+    fed = federados_filt = fed_raw[pd.to_numeric(fed_raw['periodo'], errors='coerce') == int(year)].copy()
+    
+    gas_raw = pd.read_parquet(gas_path)
+    gas = gas_filt = gas_raw[pd.to_numeric(gas_raw['periodo'], errors='coerce') == int(year)].copy()
+
+    # 2. Coalescer columnas solo en el subconjunto
+    fed = coalesce_normalized_columns(fed, {
+        "federacion": ["Federación", "Federacion"],
+        "comunidad_autonoma": ["Comunidad autónoma", "Comunidad autonoma"],
+        "total_raw": ["Total"],
+    })
+    gas = coalesce_normalized_columns(gas, {
+        "indicador": ["Indicador"],
+        "comunidad_autonoma": ["Comunidad autónoma", "Comunidad autonoma"],
+        "total_raw": ["Total"],
+    })
+
+    # 3. Procesar datos (mojibake, números, keys)
+    for df in (fed, gas):
+        if "total_raw" in df.columns:
+            df["Total_Num"] = df["total_raw"].map(parse_spanish_number)
+            
+        for text_col in [col for col in ("federacion", "comunidad_autonoma", "indicador") if col in df.columns]:
+            unique_vals = df[text_col].dropna().unique()
+            val_map = {v: repair_mojibake(v) for v in unique_vals}
+            df[text_col] = df[text_col].map(val_map)
+            
+        if "comunidad_autonoma" in df.columns:
+            unique_ccaa = df["comunidad_autonoma"].dropna().unique()
+            ccaa_map = {v: normalize_for_match(str(v)) for v in unique_ccaa}
+            df["ccaa_key"] = df["comunidad_autonoma"].map(ccaa_map)
+
+    # 4. Filtrar y Cruzar
+    fed_year = (
+        fed[
+            (fed['archivo_origen'] == 'federado_01.csv')
+            & (fed['federacion'].str.strip().str.upper() == 'TOTAL')
+            & (~fed['ccaa_key'].isin(EXCLUDED_KEYS))
+        ][['comunidad_autonoma', 'ccaa_key', 'Total_Num']]
+        .rename(columns={'Total_Num': 'Licencias Federadas', 'comunidad_autonoma': 'CCAA'})
+        .copy()
+    )
+
+    gas_year = (
+        gas[
+            (gas['archivo_origen'] == 'gasto_03.csv')
+            & (gas['indicador'].str.contains('Gasto medio por hogar', na=False, case=False))
+            & (gas['ccaa_key'] != normalize_for_match('TOTAL'))
+        ][['ccaa_key', 'Total_Num']]
+        .rename(columns={'Total_Num': 'Gasto Promedio Hogar Eur'})
+    )
+
+    merged_df = pd.merge(fed_year, gas_year, on='ccaa_key', how='inner')
+    log.info(f"✅ build_home_data({year}) ok: {len(merged_df)} filas.")
+    return merged_df.drop(columns=['ccaa_key'])
+
 # ── Helpers internos ───────────────────────────────────────────────────────────
 def _check_toxicity(text: str) -> tuple[bool, float]:
     """Devuelve (is_toxic, score)."""
@@ -240,6 +420,11 @@ def _generate(prompt: str, lang: str, max_new_tokens: int, temperature: float,
 
 
 # ── Endpoints ──────────────────────────────────────────────────────────────────
+@app.get("/")
+def read_root():
+    return {"message": "Welcome to the DEPORTEData API"}
+
+
 @app.get("/health", response_model=HealthResponse, tags=["Sistema"])
 def health():
     """Comprueba si el servidor y los modelos están listos."""
@@ -252,7 +437,97 @@ def health():
     )
 
 
-@app.post("/toxicity", response_model=ToxicityResponse, tags=["Modelos"])
+@app.post("/api/v1/token", tags=["Seguridad"])
+def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
+    user = USERS_DB.get(form_data.username)
+    if not user or user["password"] != form_data.password:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    return {"access_token": f"fake-token-{user['username']}", "token_type": "bearer"}
+
+
+@app.get("/api/v1/dashboard/metrics/{year}", tags=["Dashboard"])
+def get_dashboard_metrics(year: int, territory: str = "Todas las CCAA"):
+    try:
+        df = build_home_data(year)
+        if territory != "Todas las CCAA" and territory != "All Regions":
+            df = df[df['CCAA'] == territory]
+        
+        if df.empty:
+            raise HTTPException(status_code=404, detail="Data for this year not found")
+
+        return {
+            "avg_spending": round(float(df['Gasto Promedio Hogar Eur'].mean()), 2),
+            "total_licenses": int(df['Licencias Federadas'].sum()),
+            "areas_analyzed": len(df)
+        }
+    except FileNotFoundError as e:
+        raise HTTPException(status_code=404, detail=str(e))
+    except Exception as e:
+        log.error(f"Error in metrics: {e}")
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@app.get("/api/v1/dashboard/charts/{year}", tags=["Dashboard"])
+def get_dashboard_charts(year: int, territory: str = "Todas las CCAA"):
+    try:
+        df = build_home_data(year)
+        if territory != "Todas las CCAA" and territory != "All Regions":
+            df = df[df['CCAA'] == territory]
+            
+        if df.empty:
+            raise HTTPException(status_code=404, detail="Data for this year not found")
+
+        return df.to_dict(orient="records")
+    except FileNotFoundError as e:
+        raise HTTPException(status_code=404, detail=str(e))
+    except Exception as e:
+        log.error(f"Error in charts: {e}")
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@app.get("/api/v1/dashboard/filters", tags=["Dashboard"])
+def get_dashboard_filters():
+    return {
+        "years": [str(y) for y in range(2023, 2005, -1)],
+        "territories": [
+            "Todas las CCAA",
+            "Andalucía", "Aragón", "Asturias, Principado de", "Balears, Illes",
+            "Canarias", "Cantabria", "Castilla y León", "Castilla-La Mancha",
+            "Cataluña", "Comunitat Valenciana", "Extremadura", "Galicia",
+            "Madrid, Comunidad de", "Murcia, Región de", "Navarra, Comunidad Foral de",
+            "País Vasco", "Rioja, La"
+        ]
+    }
+
+
+@app.get("/api/v1/admin/stats", tags=["Admin"])
+def get_admin_stats(token: str = Depends(oauth2_scheme)):
+    """Requiere autenticación Bearer."""
+    get_current_user(token)
+    return {
+        "active_users": 142,
+        "total_queries": 2840,
+        "chart_q": np.random.randn(20).tolist(),
+        "chart_v": np.random.randn(20).tolist(),
+        "total_by_day": np.random.randint(10, 100, size=7).tolist(),
+        "failed_attempts": 3,
+        "logs": [
+            {"User": "admin", "IP": "192.168.1.45", "Status": "Success"},
+            {"User": "root", "IP": "85.23.11.102", "Status": "Blocked"},
+            {"User": "guest", "IP": "172.16.0.5", "Status": "Failed"},
+            {"User": "admin", "IP": "192.168.1.45", "Status": "Success"}
+        ],
+        "cpu_load": "24%",
+        "ram_usage": "1.2 GB",
+        "system_load": np.random.randn(20).tolist()
+    }
+
+
+@app.post("/toxicity", response_model=ToxicityResponse, tags=["AI"])
 def check_toxicity_endpoint(req: ToxicityRequest):
     """
     Comprueba si un texto es tóxico.
@@ -265,7 +540,7 @@ def check_toxicity_endpoint(req: ToxicityRequest):
     return ToxicityResponse(is_toxic=is_toxic, score=round(score, 4))
 
 
-@app.post("/chat", response_model=ChatResponse, tags=["Modelos"])
+@app.post("/chat", response_model=ChatResponse, tags=["AI"])
 def chat_endpoint(req: ChatRequest):
     """
     Genera una respuesta del chatbot.
