+++ dashboard/app.py
@@ -16,6 +16,32 @@ except ImportError:
         check_toxicity, load_models, generate_llm_response,
     )
 
+import requests
+import json
+
+API_URL = "http://localhost:8000"
+
+def get_api_data(endpoint, params=None, auth_token=None):
+    """Helper para llamadas GET a la API."""
+    headers = {"Authorization": f"Bearer {auth_token}"} if auth_token else {}
+    try:
+        resp = requests.get(f"{API_URL}{endpoint}", params=params, headers=headers, timeout=10)
+        if resp.status_code == 200:
+            return resp.json()
+    except Exception:
+        pass
+    return None
+
+def post_api_data(endpoint, json_data):
+    """Helper para llamadas POST a la API."""
+    try:
+        resp = requests.post(f"{API_URL}{endpoint}", json=json_data, timeout=30)
+        if resp.status_code == 200:
+            return resp.json()
+    except Exception:
+        pass
+    return None
+
 import unicodedata
 
 def normalize(text: str) -> str:
@@ -38,6 +64,8 @@ MANUAL_TOXIC_TERMS = [
 # Inicialización de estado de sesión
 if 'is_admin' not in st.session_state:
     st.session_state.is_admin = False
+if 'auth_token' not in st.session_state:
+    st.session_state.auth_token = None
 if 'show_login' not in st.session_state:
     st.session_state.show_login = False
 if 'theme' not in st.session_state:
@@ -370,166 +398,11 @@ def apply_plotly_style(fig):
     return fig
 
 
-def repair_mojibake(text: str) -> str:
-    """Repara textos UTF-8 mal decodificados como latin-1 cuando es posible."""
-    clean = str(text).replace("\ufeff", "").replace("ï»¿", "")
-    try:
-        repaired = clean.encode("latin1").decode("utf-8")
-        bad_before = clean.count("Ã") + clean.count("ï")
-        bad_after = repaired.count("Ã") + repaired.count("ï")
-        if bad_after <= bad_before:
-            clean = repaired
-    except (UnicodeEncodeError, UnicodeDecodeError):
-        pass
-    return clean.strip()
+# Eliminadas funciones de carga local (mojibake, normalize_column, build_home_data) 
+# porque ahora la lógica reside en la API unificada.
 
-
-def normalize_column_name(name: str) -> str:
-    """Normaliza nombres de columnas eliminando BOM, tildes y ruido de encoding."""
-    clean = repair_mojibake(name).lower()
-    clean = unicodedata.normalize("NFKD", clean)
-    clean = "".join(c for c in clean if not unicodedata.combining(c))
-    return clean
-
-
-def coalesce_normalized_columns(df: pd.DataFrame, canonical_map: dict[str, list[str]]) -> pd.DataFrame:
-    """Crea columnas canónicas combinando variantes equivalentes por nombre."""
-    result = df.copy()
-    normalized_lookup: dict[str, list[str]] = {}
-    for col in result.columns:
-        normalized_lookup.setdefault(normalize_column_name(col), []).append(col)
-
-    for canonical_name, aliases in canonical_map.items():
-        candidate_cols: list[str] = []
-        for alias in aliases:
-            alias_key = normalize_column_name(alias)
-            candidate_cols.extend(normalized_lookup.get(alias_key, []))
-            for normalized_col, original_cols in normalized_lookup.items():
-                if normalized_col.endswith(alias_key) and normalized_col != alias_key:
-                    candidate_cols.extend(original_cols)
-        if not candidate_cols:
-            continue
-        unique_candidates = list(dict.fromkeys(candidate_cols))
-        result[canonical_name] = result[unique_candidates].bfill(axis=1).iloc[:, 0]
-
-    return result
-
-
-def parse_spanish_number(value):
-    """Convierte números en formato español a float."""
-    if pd.isna(value):
-        return np.nan
-    text = str(value).strip().replace("\xa0", "").replace(" ", "")
-    if not text or text == "..":
-        return np.nan
-    if "," in text:
-        text = text.replace(".", "").replace(",", ".")
-    elif text.count(".") > 1:
-        text = text.replace(".", "")
-    elif text.count(".") == 1:
-        left, right = text.split(".")
-        if right.isdigit() and len(right) == 3 and left.isdigit():
-            text = left + right
-    try:
-        return float(text)
-    except ValueError:
-        return np.nan
-@st.cache_data
-def load_home_data(year: str) -> pd.DataFrame:
-    """Carga y une federados.parquet + gasto.parquet para el año indicado.
-    federados usa comas ('asturias, principado de') y gasto usa paréntesis
-    ('asturias (principado de)'), por lo que se normaliza ccaa_limpia antes del join."""
-    EXCLUDED_CCAA = {'TOTAL', 'Sin territorializar', 'Ceuta', 'Melilla'}
-
-    # Mapeo ccaa_limpia de federados → clave equivalente en gasto
-    CCAA_KEY_MAP = {
-        'asturias, principado de':     'asturias (principado de)',
-        'balears, illes':              'balears (illes)',
-        'madrid, comunidad de':        'madrid (comunidad de)',
-        'murcia, regi\u00f3n de':           'murcia (regi\u00f3n de)',
-        'navarra, comunidad foral de': 'navarra (comunidad foral de)',
-        'rioja, la':                   'rioja (la)',
-    }
-
-    fed = coalesce_normalized_columns(
-        pd.read_parquet("data/processed/federados.parquet"),
-        {
-            "federacion": ["Federación", "Federacion"],
-            "comunidad_autonoma": ["Comunidad autónoma", "Comunidad autonoma"],
-            "periodo": ["periodo"],
-            "total_raw": ["Total"],
-        },
-    )
-    gas = coalesce_normalized_columns(
-        pd.read_parquet("data/processed/gasto.parquet"),
-        {
-            "indicador": ["Indicador"],
-            "comunidad_autonoma": ["Comunidad autónoma", "Comunidad autonoma"],
-            "periodo": ["periodo"],
-            "total_raw": ["Total"],
-        },
-    )
-    yr = int(year)
-
-    for df in (fed, gas):
-        df["periodo"] = pd.to_numeric(df["periodo"], errors="coerce")
-        df["Total_Num"] = df["total_raw"].map(parse_spanish_number)
-        for text_col in [col for col in ("federacion", "comunidad_autonoma", "indicador") if col in df.columns]:
-            df[text_col] = df[text_col].map(lambda x: repair_mojibake(x) if pd.notna(x) else x)
-        df["ccaa_limpia"] = df["comunidad_autonoma"].map(
-            lambda x: normalize(str(x)) if pd.notna(x) else x
-        )
-
-    # Licencias totales por CCAA (fuente federado_01, fila 'TOTAL' de federación)
-    fed_year = (
-        fed[
-            (fed['periodo'] == yr)
-            & (fed['archivo_origen'] == 'federado_01.csv')
-            & (fed['federacion'] == 'TOTAL')
-            & (~fed['comunidad_autonoma'].isin(EXCLUDED_CCAA))
-        ][['comunidad_autonoma', 'ccaa_limpia', 'Total_Num']]
-        .rename(columns={'Total_Num': 'Licencias_Federadas', 'comunidad_autonoma': 'CCAA'})
-        .copy()
-    )
-    # Normalizar ccaa_limpia de federados para que coincida con gasto
-    fed_year['ccaa_key'] = fed_year['ccaa_limpia'].map(
-        lambda x: CCAA_KEY_MAP.get(x, x)
-    )
-
-    # Gasto medio por hogar por CCAA (fuente gasto_03)
-    gas_year = (
-        gas[
-            (gas['periodo'] == yr)
-            & (gas['archivo_origen'] == 'gasto_03.csv')
-            & (gas['indicador'] == 'Gasto medio por hogar (Euros)')
-            & (gas['comunidad_autonoma'] != 'TOTAL')
-        ][['ccaa_limpia', 'Total_Num']]
-        .rename(columns={'Total_Num': 'Gasto_Promedio_Hogar_Eur', 'ccaa_limpia': 'ccaa_key'})
-    )
-
-    df = pd.merge(fed_year, gas_year, on='ccaa_key', how='inner')
-    df = df.drop(columns=['ccaa_limpia', 'ccaa_key'])
-    return df
-
-# Título y encabezado
-st.title(L['main_title'])
-st.markdown(f"### {L['main_subtitle']}")
-
-# Pantalla de login
-if st.session_state.show_login:
-    st.header(L['login_header'])
-    with st.form("login_form"):
-        username = st.text_input(L['login_user'])
-        password = st.text_input(L['login_pass'], type="password")
-        submit = st.form_submit_button(L['login_btn'])
-        if submit:
-            if username == "admin" and password == "1234":
-                st.session_state.is_admin = True
-                st.session_state.show_login = False
-                st.rerun()
-            else:
-                st.error(L['login_err'])
-    st.stop()
+# Eliminadas funciones de carga local (mojibake, normalize_column, build_home_data) 
+# porque ahora la lógica reside en la API unificada.
 
 # Definición de pestañas
 tabs_list = [L['tab_home'], L['tab_ai']]
@@ -543,26 +416,30 @@ if st.session_state.is_admin:
 with tab1:
     st.header(f"{L['dash_header']} - {st.session_state.sel_year}")
     try:
-        # Cargar datos desde federados.parquet y gasto.parquet
-        df_real = load_home_data(st.session_state.sel_year)
-        df_display = df_real.rename(columns={
-            'Gasto_Promedio_Hogar_Eur': L['col_gasto'],
-            'Licencias_Federadas': L['col_lic'],
+        # 1. Obtener gráficos y métricas desde la API unificada
+        year = st.session_state.sel_year
+        territory = st.session_state.sel_territory
+        
+        # Obtener datos de gráficos
+        data_charts = get_api_data(f"/api/v1/dashboard/charts/{year}", params={"territory": territory})
+        # Obtener métricas agregadas
+        data_metrics = get_api_data(f"/api/v1/dashboard/metrics/{year}", params={"territory": territory})
+
+        if not data_charts:
+             raise Exception("API Return empty data")
+
+        df_filtered = pd.DataFrame(data_charts).rename(columns={
+            'Gasto Promedio Hogar Eur': L['col_gasto'],
+            'Licencias Federadas': L['col_lic'],
             'CCAA': L['col_ccaa'],
         })
 
-        # Aplicar filtro de Territorio
-        if st.session_state.sel_territory != "Todas las CCAA":
-            df_filtered = df_display[df_real['CCAA'] == st.session_state.sel_territory]
-        else:
-            df_filtered = df_display
-
-        # Métricas Dinámicas
+        # Métricas Dinámicas desde la API
         col1, col2, col3 = st.columns(3)
-        if not df_filtered.empty:
-            avg_gasto = df_filtered[L['col_gasto']].mean()
-            total_licencias = df_filtered[L['col_lic']].sum()
-            num_ccaa = len(df_filtered)
+        if data_metrics:
+            avg_gasto = data_metrics["avg_spending"]
+            total_licencias = data_metrics["total_licenses"]
+            num_ccaa = data_metrics["areas_analyzed"]
 
             col1.metric(L['metric_spending'], f"€ {avg_gasto:.0f}", None)
             col2.metric(
@@ -627,32 +504,21 @@ with tab2:
             if manually_toxic:
                 assistant_response = blocked_msg
             else:
-                # ── Layer 2: load data + try AI pipeline ──────────────────
-                try:
-                    @st.cache_data
-                    def load_assistant_data():
-                        df = pd.read_parquet("data/processed/deporte_data/anio=2023/hechos_indicadores.parquet")
-                        return prepare_assistant_data(df)
-
-                    df_rag = load_assistant_data()
-
-                    # Try AI model
-                    try:
-                        with st.spinner("🤖 Generando respuesta (esto puede tardar unos segundos)..." if st.session_state.lang == "ES" else "🤖 Generating response (this might take a few seconds)..."):
-                            toxic_clf, llm_pipeline = load_models()
-                            is_toxic, _ = check_toxicity(prompt, toxic_clf)
-                            if is_toxic:
-                                assistant_response = blocked_msg
-                            else:
-                                assistant_response = generate_llm_response(
-                                    prompt, df_rag, llm_pipeline, st.session_state.lang
-                                )
-                    except Exception as ai_err:
-                        # Show the real error so we can diagnose it
-                        assistant_response = f"❌ Error al cargar el modelo IA: {type(ai_err).__name__}: {ai_err}"
-
-                except Exception as e:
-                    assistant_response = f"{L['chat_error_data']} ({str(e)})"
+                # ── Layer 2: Llamada a la API ──────────────────────────
+                json_chat = {
+                    "prompt": prompt,
+                    "lang": st.session_state.lang
+                }
+                api_resp = post_api_data("/chat", json_chat)
+                
+                if api_resp:
+                    if api_resp.get("is_toxic"):
+                        assistant_response = blocked_msg
+                    else:
+                        assistant_response = api_resp.get("response", "No response from AI")
+                else:
+                    # Fallback si la API no responde
+                    assistant_response = "⚠️ El servidor de IA no responde. Por favor, verifica que la API está encendida."
 
             for chunk in assistant_response.split():
                 full_response += chunk + " "
@@ -664,42 +530,46 @@ with tab2:
 if st.session_state.is_admin:
     with tab_admin:
         st.header(L['admin_header'])
-        st.subheader(L['admin_usage'])
-        col_u1, col_u2 = st.columns(2)
-        with col_u1:
-            st.metric(L['admin_active'], "142", None)
-            fig_usage = px.line(pd.DataFrame(np.random.randn(20, 2), columns=[L['chart_q'], L['chart_v']]), color_discrete_sequence=[accent_color, "#FF4B4B"])
-            st.plotly_chart(apply_plotly_style(fig_usage), use_container_width=True)
-        with col_u2:
-            st.metric(L['admin_queries'], "2,840", None)
-            fig_total = px.bar(np.random.randint(10, 100, size=(7, 1)), color_discrete_sequence=[accent_color])
-            st.plotly_chart(apply_plotly_style(fig_total), use_container_width=True)
         
-        st.divider()
+        # Obtener estadísticas reales de la API
+        admin_data = get_api_data("/api/v1/admin/stats", auth_token=st.session_state.auth_token)
         
-        col_s1, col_s2 = st.columns(2)
-        with col_s1:
-            st.subheader(L['admin_security'])
-            st.metric(L['admin_failed'], "3", "-12%", delta_color="normal")
-            st.markdown(f"**{L['admin_last_logs']}**")
-            admin_logs = pd.DataFrame({
-                "User": ["admin", "root", "guest", "admin"],
-                "IP": ["192.168.1.45", "85.23.11.102", "172.16.0.5", "192.168.1.45"],
-                "Status": ["Success", "Blocked", "Failed", "Success"]
-            })
-            st.table(admin_logs)
+        if admin_data:
+            st.subheader(L['admin_usage'])
+            col_u1, col_u2 = st.columns(2)
+            with col_u1:
+                st.metric(L['admin_active'], str(admin_data["active_users"]), None)
+                fig_usage = px.line(pd.DataFrame({
+                    L['chart_q']: admin_data["chart_q"],
+                    L['chart_v']: admin_data["chart_v"]
+                }), color_discrete_sequence=[accent_color, "#FF4B4B"])
+                st.plotly_chart(apply_plotly_style(fig_usage), use_container_width=True)
+            with col_u2:
+                st.metric(L['admin_queries'], f"{admin_data['total_queries']:,}", None)
+                fig_total = px.bar(admin_data["total_by_day"], color_discrete_sequence=[accent_color])
+                st.plotly_chart(apply_plotly_style(fig_total), use_container_width=True)
             
-        with col_s2:
-            st.subheader(L['admin_telemetry'])
-            col_t1, col_t2 = st.columns(2)
-            col_t1.metric(L['admin_cpu'], "24%", "2%")
-            col_t2.metric(L['admin_ram'], "1.2 GB", "0.1 GB")
-            st.markdown(f"**{L['admin_system_load']}**")
-            telemetry_data = pd.DataFrame(np.random.randn(20, 1), columns=[L['chart_l']])
-            fig_telemetry = px.area(telemetry_data, color_discrete_sequence=[accent_color])
-            st.plotly_chart(apply_plotly_style(fig_telemetry), use_container_width=True)
+            st.divider()
+            
+            col_s1, col_s2 = st.columns(2)
+            with col_s1:
+                st.subheader(L['admin_security'])
+                st.metric(L['admin_failed'], str(admin_data["failed_attempts"]), None)
+                st.markdown(f"**{L['admin_last_logs']}**")
+                st.table(pd.DataFrame(admin_data["logs"]))
+                
+            with col_s2:
+                st.subheader(L['admin_telemetry'])
+                col_t1, col_t2 = st.columns(2)
+                col_t1.metric(L['admin_cpu'], admin_data["cpu_load"], None)
+                col_t2.metric(L['admin_ram'], admin_data["ram_usage"], None)
+                st.markdown(f"**{L['admin_system_load']}**")
+                fig_telemetry = px.area(admin_data["system_load"], color_discrete_sequence=[accent_color])
+                st.plotly_chart(apply_plotly_style(fig_telemetry), use_container_width=True)
+        else:
+            st.error("No se pudo conectar con el servicio administrativo de la API.")
 
-# Sidebar y personalización
+# Sidebar y personalización (Definidos al inicio para que los filtros afecten al resto del script)
 with st.sidebar:
     st.image(icon_path, width=100)
     st.markdown("## DEPORTEData")
@@ -729,18 +599,29 @@ with st.sidebar:
     st.divider()
     
     st.markdown(f"### {L['sidebar_filters']}")
-    # Filtro de Año persistente
-    # Años disponibles: intersección de federados (2005-2024) y gasto (2006-2023)
+    
+    # Callbacks para actualización instantánea
+    def on_year_change():
+        st.session_state.sel_year = st.session_state.year_selector
+        
+    def on_territory_change():
+        val = st.session_state.territory_selector
+        if val == L['all_ccaa']:
+            st.session_state.sel_territory = "Todas las CCAA"
+        else:
+            st.session_state.sel_territory = val
+
+    # Filtro de Año
     year_options = [str(y) for y in range(2023, 2005, -1)]
-    sel_year_idx = year_options.index(st.session_state.sel_year) if st.session_state.sel_year in year_options else 0
     st.selectbox(
         L['filter_year'],
         year_options,
-        index=sel_year_idx,
-        key='sel_year'
+        index=year_options.index(st.session_state.sel_year) if st.session_state.sel_year in year_options else 0,
+        key='year_selector',
+        on_change=on_year_change
     )
 
-    # Filtro de Territorio persistente (nombres exactos de federados.parquet)
+    # Filtro de Territorio
     territory_options = [
         L['all_ccaa'],
         "Andalucía", "Aragón", "Asturias, Principado de", "Balears, Illes",
@@ -749,17 +630,18 @@ with st.sidebar:
         "Madrid, Comunidad de", "Murcia, Región de", "Navarra, Comunidad Foral de",
         "País Vasco", "Rioja, La",
     ]
+    
+    # Determinar el índice actual basado en el estado (traduciendo "Todas las CCAA" si es necesario)
+    current_territory = st.session_state.sel_territory
+    display_territory = L['all_ccaa'] if current_territory == "Todas las CCAA" else current_territory
+    
     st.selectbox(
         L['filter_territory'], 
         territory_options, 
-        index=territory_options.index(st.session_state.sel_territory) if st.session_state.sel_territory in territory_options else 0,
-        key='sel_territory_input'
+        index=territory_options.index(display_territory) if display_territory in territory_options else 0,
+        key='territory_selector',
+        on_change=on_territory_change
     )
-    # Sincronizar sel_territory con el nombre interno (Todas las CCAA)
-    if st.session_state.sel_territory_input == L['all_ccaa']:
-        st.session_state.sel_territory = "Todas las CCAA"
-    else:
-        st.session_state.sel_territory = st.session_state.sel_territory_input
 
     st.divider()
     
@@ -769,3 +651,30 @@ with st.sidebar:
             st.rerun()
     else:
         st.success(L['admin_label'])
+
+# --- Títulos y Pestañas ---
+st.title(L['main_title'])
+st.markdown(f"### {L['main_subtitle']}")
+
+# Pantalla de login
+if st.session_state.show_login:
+    st.header(L['login_header'])
+    with st.form("login_form"):
+        username = st.text_input(L['login_user'])
+        password = st.text_input(L['login_pass'], type="password")
+        submit = st.form_submit_button(L['login_btn'])
+        if submit:
+            # Login contra la API para obtener Token
+            login_resp = requests.post(
+                f"{API_URL}/api/v1/token",
+                data={"username": username, "password": password}
+            )
+            if login_resp.status_code == 200:
+                data = login_resp.json()
+                st.session_state.is_admin = True
+                st.session_state.auth_token = data["access_token"]
+                st.session_state.show_login = False
+                st.rerun()
+            else:
+                st.error(L['login_err'])
+    st.stop()
