+++ requirements.txt
@@ -7,7 +7,8 @@ plotly>=5.18.0
 # ─── API local de Modelos IA ──────────────────────────────────────────────────
 fastapi>=0.111.0
 uvicorn[standard]>=0.29.0
-requests>=2.31.0
+requests
+python-multipart>=2.31.0
 
 # ─── Deep Learning base ──────────────────────────────────────────────────────
 torch
